Rename field_path implementation modules
