+++ docs/reference/field_path.md
@@ -22,7 +22,7 @@ def read_field_paths(r: BitReader) -> list[FieldPath]
 
 Decode a Huffman-coded sequence of field paths from r.
 
-Source: [src/gem/schema/field_path/decoder.py:20](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/decoder.py#L20)
+Source: [src/gem/schema/field_path/path_sequence.py:20](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/path_sequence.py#L20)
 
 ## `gem.schema.field_path.FieldPath`
 
@@ -34,7 +34,7 @@ class FieldPath
 
 A mutable path of up to 7 integer field indices.
 
-Source: [src/gem/schema/field_path/model.py:8](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/model.py#L8)
+Source: [src/gem/schema/field_path/models.py:8](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/models.py#L8)
 
 #### Methods
 
@@ -44,7 +44,7 @@ Signature: `def FieldPath.reset(self) -> None`
 
 Reset to the initial empty state.
 
-Source: [src/gem/schema/field_path/model.py:27](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/model.py#L27)
+Source: [src/gem/schema/field_path/models.py:27](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/models.py#L27)
 
 ##### `pop`
 
@@ -52,7 +52,7 @@ Signature: `def FieldPath.pop(self, n: int) -> None`
 
 Pop n levels off the path, zeroing the vacated slots.
 
-Source: [src/gem/schema/field_path/model.py:33](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/model.py#L33)
+Source: [src/gem/schema/field_path/models.py:33](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/models.py#L33)
 
 ##### `copy`
 
@@ -60,7 +60,7 @@ Signature: `def FieldPath.copy(self) -> FieldPath`
 
 Return an independent copy of this path.
 
-Source: [src/gem/schema/field_path/model.py:43](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/model.py#L43)
+Source: [src/gem/schema/field_path/models.py:43](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/models.py#L43)
 
 ##### `to_tuple`
 
@@ -68,7 +68,7 @@ Signature: `def FieldPath.to_tuple(self) -> tuple[int, ...]`
 
 Return the active indices as an immutable tuple.
 
-Source: [src/gem/schema/field_path/model.py:55](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/model.py#L55)
+Source: [src/gem/schema/field_path/models.py:55](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/models.py#L55)
 
 ##### `to_str`
 
@@ -76,7 +76,7 @@ Signature: `def FieldPath.to_str(self) -> str`
 
 Return a slash-separated string of active indices.
 
-Source: [src/gem/schema/field_path/model.py:63](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/model.py#L63)
+Source: [src/gem/schema/field_path/models.py:63](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/models.py#L63)
 
 ##### `plus_one`
 
@@ -84,7 +84,7 @@ Signature: `def FieldPath.plus_one(self) -> None`
 
 Increment the deepest index by 1.
 
-Source: [src/gem/schema/field_path/model.py:71](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/model.py#L71)
+Source: [src/gem/schema/field_path/models.py:71](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/models.py#L71)
 
 ## `gem.schema.field_path.FieldPathOp`
 
@@ -96,7 +96,7 @@ class FieldPathOp
 
 A single field-path operation with its Huffman weight.
 
-Source: [src/gem/schema/field_path/ops.py:20](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/ops.py#L20)
+Source: [src/gem/schema/field_path/operations.py:20](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/schema/field_path/operations.py#L20)
 
 #### Dataclass fields
